target files may not exist initially

+++ deploy.sh
@@ -7,7 +7,7 @@ for file in $(ls -Ad $HOME/.dotfiles/.[a-z]*); do
 	if [ -f $file ]; then
 		target=$HOME/$(basename $file)
 		echo "aliasing file $file to $target" 
-		mv $target $target.bak
+		[ -f $target ] && mv $target $target.bak
 		ln -sf $file $target
 	fi
